refactor(abr): move flip_amount_for_credit_card next to its only caller
The helper was previously defined in overrides/bank_transaction.py because
it served as a before_insert doc_event handler on Bank Transaction. Since
the previous commit scoped the flip to the bank statement importer path
only, the function no longer belongs alongside ExtendedBankTransaction.

Move it into bank_statement_importer.py (its sole call site) and drop the
cross-module import. No behavior change.
